+++ addOns/spiderAjax/src/test/java/org/zaproxy/zap/extension/spiderAjax/AjaxSpiderParamUnitTest.java
@@ -21,6 +21,7 @@ package org.zaproxy.zap.extension.spiderAjax;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.contains;
+import static org.hamcrest.Matchers.empty;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.is;
 import static org.mockito.Mockito.mock;
@@ -69,9 +70,10 @@ class AjaxSpiderParamUnitTest {
 
     @ParameterizedTest
     @NullSource
-    @ValueSource(ints = {1, 2, 3, 4})
+    @ValueSource(ints = {1, 2, 3, 4, 5})
     void shouldHaveAllowedResourcesByDefault(Integer version) {
         // Given
+        configuration = new ZapXmlConfiguration();
         configuration.setProperty(param.getConfigVersionKey(), version);
         // When
         param.load(configuration);
@@ -81,6 +83,32 @@ class AjaxSpiderParamUnitTest {
                 contains(
                         allowedResource("^http.*\\.js(?:\\?.*)?$"),
                         allowedResource("^http.*\\.css(?:\\?.*)?$")));
+    }
+
+    @Test
+    void shouldNotAddDefaultAllowedResourcesForVersion5WithExistingResources() {
+        // Given
+        configuration = new ZapXmlConfiguration();
+        configuration.setProperty(param.getConfigVersionKey(), 5);
+        String allowedResourceKey = "ajaxSpider.allowedResources.allowedResource(0).";
+        String regex = "^https?://example\\.com/.*";
+        configuration.setProperty(allowedResourceKey + "regex", regex);
+        configuration.setProperty(allowedResourceKey + "enabled", "true");
+        // When
+        param.load(configuration);
+        // Then
+        assertThat(param.getAllowedResources(), contains(allowedResource(regex)));
+    }
+
+    @Test
+    void shouldNotAddDefaultAllowedResourcesForVersion6() {
+        // Given
+        configuration = new ZapXmlConfiguration();
+        configuration.setProperty(param.getConfigVersionKey(), 6);
+        // When
+        param.load(configuration);
+        // Then
+        assertThat(param.getAllowedResources(), is(empty()));
     }
 
     @ParameterizedTest